removed unused imports

+++ src/net/minecraft/server/v1_8_R3/ControllerLook.java
@@ -17,7 +17,7 @@ public class ControllerLook {
     public void a(Entity entity, float f, float f1) {
         this.e = entity.locX;
         if (entity instanceof EntityLiving) {
-            this.f = entity.locY + (double) entity.getHeadHeight();
+            this.f = entity.locY + entity.getHeadHeight();
         } else {
             this.f = (entity.getBoundingBox().b + entity.getBoundingBox().e) / 2.0D;
         }
@@ -42,9 +42,9 @@ public class ControllerLook {
         if (this.d) {
             this.d = false;
             double d0 = this.e - this.a.locX;
-            double d1 = this.f - (this.a.locY + (double) this.a.getHeadHeight());
+            double d1 = this.f - (this.a.locY + this.a.getHeadHeight());
             double d2 = this.g - this.a.locZ;
-            double d3 = (double) MathHelper.sqrt(d0 * d0 + d2 * d2);
+            double d3 = MathHelper.sqrt(d0 * d0 + d2 * d2);
             float f = (float) (MathHelper.b(d2, d0) * 180.0D / 3.1415927410125732D) - 90.0F;
             float f1 = (float) (-(MathHelper.b(d1, d3) * 180.0D / 3.1415927410125732D));
