+++ src/net/minecraft/server/v1_8_R3/WorldBorder.java
@@ -31,7 +31,7 @@ public class WorldBorder {
     }
 
     public boolean a(BlockPosition blockposition) {
-        return (double) (blockposition.getX() + 1) > this.b() && (double) blockposition.getX() < this.d() && (double) (blockposition.getZ() + 1) > this.c() && (double) blockposition.getZ() < this.e();
+        return blockposition.getX() + 1 > this.b() && blockposition.getX() < this.d() && blockposition.getZ() + 1 > this.c() && blockposition.getZ() < this.e();
     }
 
     // CraftBukkit start - split method
@@ -46,7 +46,7 @@ public class WorldBorder {
 
     // Inlined the getters from ChunkCoordIntPair
     public boolean isInBounds(int x, int z) {
-        return (double) ((x  << 4) + 15) > this.b() && (double) (x << 4) < this.d() && (double) ((z  << 4) + 15) > this.c() && (double) (x << 4) < this.e();
+        return (x  << 4) + 15 > this.b() && x << 4 < this.d() && (z  << 4) + 15 > this.c() && x << 4 < this.e();
     }
 
     public boolean a(AxisAlignedBB axisalignedbb) {
@@ -75,8 +75,8 @@ public class WorldBorder {
     public double b() {
         double d0 = this.getCenterX() - this.getSize() / 2.0D;
 
-        if (d0 < (double) (-this.h)) {
-            d0 = (double) (-this.h);
+        if (d0 < (-this.h)) {
+            d0 = (-this.h);
         }
 
         return d0;
@@ -85,8 +85,8 @@ public class WorldBorder {
     public double c() {
         double d0 = this.getCenterZ() - this.getSize() / 2.0D;
 
-        if (d0 < (double) (-this.h)) {
-            d0 = (double) (-this.h);
+        if (d0 < (-this.h)) {
+            d0 = (-this.h);
         }
 
         return d0;
@@ -95,8 +95,8 @@ public class WorldBorder {
     public double d() {
         double d0 = this.getCenterX() + this.getSize() / 2.0D;
 
-        if (d0 > (double) this.h) {
-            d0 = (double) this.h;
+        if (d0 > this.h) {
+            d0 = this.h;
         }
 
         return d0;
@@ -105,8 +105,8 @@ public class WorldBorder {
     public double e() {
         double d0 = this.getCenterZ() + this.getSize() / 2.0D;
 
-        if (d0 > (double) this.h) {
-            d0 = (double) this.h;
+        if (d0 > this.h) {
+            d0 = this.h;
         }
 
         return d0;
@@ -135,7 +135,7 @@ public class WorldBorder {
 
     public double getSize() {
         if (this.getState() != EnumWorldBorderState.STATIONARY) {
-            double d0 = (double) ((float) (System.currentTimeMillis() - this.g) / (float) (this.f - this.g));
+            double d0 = (float) (System.currentTimeMillis() - this.g) / (float) (this.f - this.g);
 
             if (d0 < 1.0D) {
                 return this.d + (this.e - this.d) * d0;
